// src/types/enrichment.ts
export interface EnrichmentData {
  website?: string;
  phone?: string;
  type?: string;
  openingHours?: {
    [key: string]: string;
  };
  facebook?: string;
  instagram?: string;
  tiktok?: string;
  snapchat?: string;
  twitter?: string;
  linkedin?: string;
  github?: string;
  youtube?: string;
  pinterest?: string;
  email_1?: string;
  email_2?: string;
}